added number of topic, latest offset and message behind value in to consumer group ui

+++ src/main/java/com/thanhpham/Kafka/dto/uiformat/ConsumerGroupDetailUI.java
@@ -11,8 +11,8 @@ public class ConsumerGroupDetailUI {
     private String consumerGroupId;
     private int memberNum;
     private int topicNum;
-    private int latestOffset;
-    private int messageBehind;
+    private long latestOffset;
+    private long messageBehind;
     private int coordinator;
     private String state;
 }
